refactor(dir): ♻ map lang surrogate key, convert all surrogate keys to longs

+++ services/Archival/FilterLists.Archival.Domain/ListArchives/ListArchive.cs
@@ -1,3 +1,3 @@
 ﻿namespace FilterLists.Archival.Domain.ListArchives;
 
-public record ListArchive(int Id, IAsyncEnumerable<ListArchiveSegment> Segments);
+public record ListArchive(long Id, IAsyncEnumerable<ListArchiveSegment> Segments);
