+++ services/Directory/FilterLists.Directory.Infrastructure/Persistence/Queries/Entities/FilterListMaintainer.cs
@@ -7,9 +7,9 @@ namespace FilterLists.Directory.Infrastructure.Persistence.Queries.Entities;
 
 public record FilterListMaintainer
 {
-    public int FilterListId { get; init; }
+    public long FilterListId { get; init; }
     public FilterList FilterList { get; init; } = null!;
-    public int MaintainerId { get; init; }
+    public long MaintainerId { get; init; }
     public Maintainer Maintainer { get; init; } = null!;
 }
